Finally solved the following nasty SML type error:
  semant.sml:70.11-78.16 Error: right-hand-side of clause doesn't agree with function result type [tycon mismatch]
    expression:  {trexp:'Z, ty:Types.ty}
    result type:  {exp:unit, ty:Types.ty}

The following line:

    let val {trexp, ty} = transExp(venv, tenv, init)
             *****

needed to be:

    let val {exp=trexp, ty} = transExp(venv, tenv, init)
             *********

I was trying to bind the first field of the result record to a local variable called 'trexp'. However this notation is not allow (as the declared name of the field is 'exp'). This caused the compiler to assume that I was defined a _new_ type :(.